backend/requirements.txt
fastapi
uvicorn==0.30.1
python-dotenv
pinecone-client
langchain-google-genai
langchain-pinecone==0.1.3
langchain==0.2.0
langchain-core==0.2.10
python-multipart
langchain-text-splitters==0.2.2
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
uvloop==0.19.0
lxml
SQLAlchemy
psycopg2-binary
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
email-validator
pandas
selenium
playwright
opencv-python
openai
ffmpeg-python
cloudinary
# Contact Enricher dependencies
facebook-sdk
phonenumbers
aiohttp 
moviepy 
reportlab
httpx
# Rebuild triggered at 2025-07-23T13:46:23.332901
boto3